fix: changed tag snippet plugin to use inject lifecycle method
